refactor - move the ux where you actually format the flow to its own file/component
